+++ filter.php
@@ -446,11 +446,15 @@ class filter_filtercodes extends moodle_text_filter {
                         // If the tag exists and is not set to "Not visible" in the custom profile field's settings.
                         if (stripos($text, '{profile_field_' . $field->shortname . '}') !== false && $field->visible != '0') {
                             $data = !empty($profiledata[$field->id]) ? $profiledata[$field->id] : '';
-                            switch ($field->datatype) {
+                            switch ($field->datatype) { // Format data for some field types.
                                 case 'datetime':
                                     // Include date and time or just date?
-                                    $datetimeformat = (!empty($field->param3) ? 'strftimedaydatetime' : 'strftimedate');
+                                    $datetimeformat = !empty($field->param3) ? 'strftimedaydatetime' : 'strftimedate';
                                     $data = empty($data) ? '' : userdate($data, get_string($datetimeformat, 'langconfig'));
+                                    break;
+                                case 'checkbox':
+                                    // 1 = Yes, 0 = No
+                                    $data = empty($data) ? get_string('no') : get_string('yes');
                                     break;
                             }
                             $replace['/\{profile_field_' . $field->shortname . '\}/i'] = $data;